pyproject.toml
[project]
name = "carolins-kasse"
version = "0.1.0"
description = "Eine Kasse mit scanner mit pygame für raspi zero"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "numpy>=2.4.2",
    "pillow>=12.1.0",
    "pygame-ce>=2.5.6",
    "python-barcode>=0.16.1",
    "reportlab>=4.4.9",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]
